Twitter4jのバージョンアップ，及びtwitter4j仕様変更への対応． ダイレクトメッセージが表示，送信できなかった問題も解決 refs #24963
git-svn-id: http://svn.sourceforge.jp/svnroot/nt-manager/NishioTweetManager/trunk@111 d8c9ecd3-d47d-4367-8645-de82c00e513f

+++ src/twitter/manage/StatusXMLConverter.java
@@ -97,7 +97,7 @@ public class StatusXMLConverter {
 		long id = status.getId();
 		String replyToScreenName = status.getInReplyToScreenName();
 		long replyStatusID = status.getInReplyToStatusId();
-		int replyUserID = status.getInReplyToUserId();
+		long replyUserID = status.getInReplyToUserId();
 		String place = null;
 		if (status.getPlace() != null) {
 			place = status.getPlace().toString();
